Refactoring: move part of util.h to stringutil.h
