+++ src/part_finder.cpp
@@ -1,4 +1,5 @@
 #include "part_finder.h"
+#include "helpers/stringutil.h"
 #include "helpers/util.h"
 
 namespace bricksim::part_finder {
@@ -23,9 +24,9 @@ namespace bricksim::part_finder {
         expression(std::move(expression)) {}
 
     bool Predicate::matches(const ldr::File& part) const {
-        return (util::containsIgnoreCase(part.metaInfo.title, expression)
-                || util::containsIgnoreCase(part.metaInfo.name, expression)
-                || util::containsIgnoreCase(part.metaInfo.theme, expression)
-                || std::any_of(part.metaInfo.keywords.begin(), part.metaInfo.keywords.end(), [this](const auto& keyword) { return util::containsIgnoreCase(keyword, expression); }));
+        return (stringutil::containsIgnoreCase(part.metaInfo.title, expression)
+                || stringutil::containsIgnoreCase(part.metaInfo.name, expression)
+                || stringutil::containsIgnoreCase(part.metaInfo.theme, expression)
+                || std::any_of(part.metaInfo.keywords.begin(), part.metaInfo.keywords.end(), [this](const auto& keyword) { return stringutil::containsIgnoreCase(keyword, expression); }));
     }
 }
