+++ src/test/helpers/CMakeLists.txt
@@ -3,5 +3,6 @@ target_sources(BrickSimTests PRIVATE
         test_fraction.cpp
         test_geometry.cpp
         test_polygon_cliping.cpp
+        test_stringutil.cpp
         test_util.cpp
         )
